requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cssselect==1.2.0
defusedxml==0.7.1
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.6
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.0.1
zope.interface==5.5.1
